\name{coeftest}
\alias{coeftest}
\title{Extract summary statistics}
\description{Extract information from poisson regression model objects of class \code{\link{cold}}.}
\usage{coeftest(object)}
\arguments{
\item{object}{an object of class \code{\link[=cold-class]{cold}}.}}
\value{Extract a list of summary statistics from poisson regression model corresponding to the fixed effects coefficients.}
\keyword{function}
