\name{cold-internal}
\alias{gintp0}
\alias{gintp}
\alias{intp0}
\alias{intp}
\alias{pssgrd}
\alias{pssgrd0}
\alias{psslik}
\alias{psslik0}
\alias{gradlogL.pss0}
\alias{gradLogL.pss0I}
\alias{gradLogL.pss0Ic}
\alias{gradLogL.pss0Ic2}
\alias{gradLogL.pss0MC}
\alias{gradLogL.pss0MC2}
\alias{gradLogL.pss1}
\alias{gradLogL.pss1I}
\alias{gradLogL.pss1Ic}
\alias{gradLogL.pss1Ic2}
\alias{gradLogL.pssMC}
\alias{gradLogL.pssMC2}
\alias{logL.pss0}
\alias{LogL.pss0I}
\alias{LogL.pss0I.aux}
\alias{LogL.pss0Ic}
\alias{LogL.pss0Ic2}
\alias{LogL.pss0Ic2.aux}
\alias{LogL.pss0MC}
\alias{LogL.pss0MC2}
\alias{LogL.pss1}
\alias{LogL.pss1I}
\alias{LogL.pss1I.aux}
\alias{LogL.pss1Ic}
\alias{LogL.pss1Ic2}
\alias{LogL.pss1Ic2.aux}
\alias{LogL.pssMC1}
\alias{LogL.pssMC2}
\alias{num.info}
\alias{num.infoI}
\alias{num.infoIc}
\alias{num.infoMC}
\title{Internal functions}
\description{Functions for internal usage only.}

\keyword{function}
