\name{getAIC}
\alias{getAIC}
\title{Extract the Akaike Information Criterion}
\description{Methods for function \code{getAIC} extracting the Akaike information criterion 
for one fitted model object from class \code{\link{cold}}. }
\usage{getAIC(object)}
\arguments{
\item{object}{an object of class \code{\link[=cold-class]{cold}}.}}
\value{Returns a numeric value corresponding to the AIC of the fitted model.}

\keyword{function}
