\name{getcoef}
\alias{getcoef}
\title{Extract the coefficient estimates}
\description{Methods for function \code{getcoef} extracting the coefficient estimates of the fitted model object from class \code{\link{cold}}. }
\usage{getcoef(object)}
\arguments{
\item{object}{an object of class \code{\link[=cold-class]{cold}}.}}
\value{Returns the coefficient estimates of the fitted model.}
\keyword{function}
