\name{getLogLik}
\alias{getLogLik}
\title{Extract Log-Likelihood}
\description{Extract the Log-Likelihood of a fitted model object from class \code{\link[=cold-class]{cold}}.}
\usage{getLogLik(object)
}
\arguments{
\item{object}{an object of class \code{\link[=cold-class]{cold}}.}}
\value{Returns a numeric value corresponding to the log-Likelihood of the fitted model.}
\keyword{function}

