\name{getvcov}
\alias{getvcov}
\title{Extract variance-covariance matrix}
\description{Extract the variance-covariance matrix of a fitted model object from class \code{\link[=cold-class]{cold}}.}
\usage{getvcov(object)}
\arguments{
\item{object}{an object of class \code{\link[=cold-class]{cold}}.}}
\value{Returns a numeric value corresponding to the variance-covariance matrix.}
\keyword{function}
