\name{model.mat}
\alias{model.mat}
\title{Extract the fixed effects model matrix}
\description{Methods for function \code{model.mat} extracting the fixed effects model matrix for a fitted model object from class \code{\link[=cold-class]{cold}}.}
\usage{model.mat(object)}
\arguments{
\item{object}{an object of class \code{\link[=cold-class]{cold}}.}}
\value{Returns a numeric value corresponding to the fixed effects model matrix.}
\keyword{function}
