\name{predict-methods}
\docType{methods}
\alias{predict-methods}
\alias{predict,cold-method}
\title{Extract "cold" Predict Values}
\description{ Methods for function \code{predict} extracting predict values of a fitted model object from class \code{\link[=cold-class]{cold}}. }
\usage{
\S4method{predict}{cold}(object)
}
\arguments{ 
  \item{object}{an object of class \code{\link[=cold-class]{cold}}.}
    }  

\section{Methods}{
\describe{
\item{\code{signature(object="cold")}:}{predict for \code{\link{cold}} object.} 
}}

\keyword{methods}

